Implementation of Connection between App and Backend Login - Done Get Recipes when logging in - Done Create Recipe - DONE Delete Recipe - TODO Logout - DONE, FIXED Register - DONE
ISSUE: When logging in, the Recipes get loaded into View Model, but when logging out they are not deleted yet, need to fix because if another user gets logged in, the old recipes are also displayed! -> RESOLVED

ISSUE: At update Recipe: MalformedJsonException Request cannot be sent! (Waiting for Nicki to finish input)